requirements.txt
python-dotenv==1.0.0
faunadb==4.5.0
cloudinary==1.32.0
pyTelegramBotAPI==4.10.0
pandas==1.5.3
requests==2.28.2
urllib3==1.26.15
aiohttp==3.8.4
asyncio==3.4.3
DateTime==5.1
jsonschema==4.17.3